Initial (partial) migration to Cucumber 7

+++ serenity-model/src/main/java/net/thucydides/core/requirements/model/cucumber/RowResultIcon.java
@@ -1,5 +1,5 @@
 package net.thucydides.core.requirements.model.cucumber;
 
 public interface RowResultIcon {
-    String resultToken(int lineNumber);
+    String resultToken(long lineNumber);
 }
